requirements.txt
asgiref==3.2.7
boto3==1.13.24
botocore==1.16.24
certifi==2020.4.5.1
chardet==3.0.4
comtypes==1.1.7
configobj==5.0.6
configparser==5.0.0
dj-database-url==0.5.0
Django==3.0.6
django-background-tasks==1.2.5
django-compat==1.0.15
django-crispy-forms==1.9.0
django-heroku==0.3.1
django-storages==1.9.1
docutils==0.15.2
docx==0.2.4
docx2pdf==0.1.7
filelock==3.0.12
gunicorn==20.0.4
httplib2==0.18.1
idna==2.9
jmespath==0.10.0
lxml==4.5.1
nibabel==3.1.0
numpy==1.18.4
packaging==20.4
pandas==1.0.4
Pillow==7.1.2
psycopg2==2.8.5
pydot==1.4.1
PyMuPDF==1.17.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-docx==0.8.10
pytz==2020.1
pyxnat==1.3
requests==2.23.0
s3transfer==0.3.3
scipy==1.4.1
six==1.15.0
sqlparse==0.3.1
tqdm==4.46.1
urllib3==1.25.9
whitenoise==5.1.0

